% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate.R
\name{as_pref}
\alias{as_pref}
\title{Coerce an object to a jpcity_pref object}
\usage{
as_pref(x)
}
\arguments{
\item{x}{A \code{integer} vector, a \code{jpcity_city} object or a \code{jpcity_pref} object.}
}
\value{
A \code{jpcity_pref} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
