% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-verb.R
\name{city_code}
\alias{city_code}
\title{Get city codes}
\usage{
city_code(city)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}
}
\value{
A \code{character} vector of city codes.
}
\description{
Get city codes
}
