% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_convert-legacy.R
\name{city_convert_legacy}
\alias{city_convert_legacy}
\title{Convert to cities at different points in time}
\usage{
city_convert_legacy(city, from, to)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}

\item{from}{A \code{character} (year, month, and day components) or date-time
object of the starting date.}

\item{to}{A \code{character} (year, month, and day components) or date-time
object of the ending date.}
}
\value{
A list of a \code{jpcity_city} object.
}
\description{
Convert to cities at different points in time
}
\examples{
city <- parse_city_legacy(c("01201", "01202"))
city_convert_legacy(city,
             from = "1970-04-01",
             to = "2020-01-01")

}
