% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-legacy.R
\name{city_data_legacy}
\alias{city_data_legacy}
\title{Get city data}
\usage{
city_data_legacy(city)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}
}
\value{
A data frame.
}
\description{
Get city data
}
