% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_desig.R
\name{city_desig_merge}
\alias{city_desig_merge}
\title{Merge designated city wards}
\usage{
city_desig_merge(city, merge_tokyo = FALSE)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}

\item{merge_tokyo}{Whether to merge Tokyo special wards?}
}
\value{
A \code{jpcity_city} object.
}
\description{
Merge designated city wards
}
\examples{
city <- parse_city(c("01101", "13101"))
city_desig_merge(city)
city_desig_merge(city,
                 merge_tokyo = TRUE)

}
