% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_desig.R
\name{city_desig_split}
\alias{city_desig_split}
\title{Split designated cities into wards}
\usage{
city_desig_split(city, split_tokyo = TRUE)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}

\item{split_tokyo}{Whether to split into Tokyo special wards?}
}
\value{
A list of a \code{jpcity_city} object.
}
\description{
Split designated cities into wards
}
\examples{
city <- parse_city(c("01100", "13100"))
city_desig_split(city)
city_desig_split(city,
                 split_tokyo = FALSE)

}
