% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-verb-legacy.R
\name{city_interval_legacy}
\alias{city_interval_legacy}
\title{Get city duration}
\usage{
city_interval_legacy(city, intersect = FALSE)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}

\item{intersect}{Whether to get the common part of the duration of cities.}
}
\value{
A \code{interval} vector of the duration of cities.
}
\description{
Get city duration
}
