% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-verb.R
\name{city_name}
\alias{city_name}
\title{Get city names}
\usage{
city_name(city, type = c("city_desig", "city"), sep = "")
}
\arguments{
\item{city}{A \code{jpcity_city} object.}

\item{type}{Types of city names. By default, returns both designated city
names (\code{"city_desig"}) and city names (\code{"city"}).}

\item{sep}{Separator for city names.}
}
\value{
A \code{character} vector of city names.
}
\description{
Get city names
}
