% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-verb-legacy.R
\name{city_name_legacy}
\alias{city_name_legacy}
\title{Get city names}
\usage{
city_name_legacy(city, type = c("city_desig", "city"), sep = "", kana = FALSE)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}

\item{type}{Types of city names. By default, returns both designated city
names (\code{"city_desig"}) and city names (\code{"city"}).}

\item{sep}{Separator for city names.}

\item{kana}{Whether to use hiragana or not?}
}
\value{
A \code{character} vector of city names.
}
\description{
Get city names
}
