% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-verb-legacy.R
\name{city_to_pref_legacy}
\alias{city_to_pref_legacy}
\title{Convert city to prefecture}
\usage{
city_to_pref_legacy(city)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}
}
\value{
A \code{jpcity_pref} object.
}
\description{
Convert city to prefecture
}
