% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{find_city}
\alias{find_city}
\title{Find cities by string patterns}
\usage{
find_city(patterns, when = NULL)
}
\arguments{
\item{patterns}{Patterns to look for. If multiple patterns are given, find
the cities that match all patterns.}

\item{when}{A \code{character} (year, month, and day components) or date-time
object.}
}
\value{
A \code{jpcity_city} object.
}
\description{
Find cities by string patterns
}
