% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_city.R
\name{get_city}
\alias{get_city}
\title{Get cities at a specific point in time}
\usage{
get_city(when)
}
\arguments{
\item{when}{A \code{character} (year, month, and day components) or date-time
object.}
}
\value{
A \code{jpcity_city} object.
}
\description{
Get cities at a specific point in time
}
\examples{
get_city("2020-01-01")

}
