% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city.R
\name{is_city}
\alias{is_city}
\title{Test if the object is a city}
\usage{
is_city(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{jpcity_city} class.
}
\description{
Test if the object is a city
}
