% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city_desig-legacy.R
\name{is_city_desig_legacy}
\alias{is_city_desig_legacy}
\title{Check if a city is a designated city or a ward of a designated city}
\usage{
is_city_desig_legacy(x, type = "city")
}
\arguments{
\item{x}{A \code{jpcity_city} object.}

\item{type}{A character vector of city types, \code{"city"} or \code{"ward"}.
By default, \code{"city"}.}
}
\value{
A logical vector.
}
\description{
Check if a city is a designated city or a ward of a designated city
}
