% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/city-legacy.R
\name{is_city_legacy}
\alias{is_city_legacy}
\title{Test if the object is a jpcity_city object}
\usage{
is_city_legacy(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{jpcity_city} class.
}
\description{
Test if the object is a jpcity_city object
}
