% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref-legacy.R
\name{is_pref_legacy}
\alias{is_pref_legacy}
\title{Test if the object is a jpcity_pref object}
\usage{
is_pref_legacy(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{jpcity_pref} class.
}
\description{
Test if the object is a jpcity_pref object
}
