% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_city.R
\name{parse_city}
\alias{parse_city}
\title{Parse city codes}
\usage{
parse_city(city_code, when = NULL, na = c("", "NA"))
}
\arguments{
\item{city_code}{A \code{character} vector of city codes.}

\item{when}{A \code{character} (year, month, and day components) or date-time
object.}

\item{na}{A \code{character} vector to be treated as missing values.}
}
\value{
A \code{jpcity_city} object.
}
\description{
Parse city codes
}
