% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pref.R
\name{parse_pref}
\alias{parse_pref}
\title{Parse prefecture codes or names}
\usage{
parse_pref(x, strict = TRUE, pref_name = deprecated())
}
\arguments{
\item{x}{A \code{character} vector of prefecture codes or names.}

\item{strict}{A scalar logical. Whether to require the code to have 1 or 2
digits. By default, \code{TRUE}.}

\item{pref_name}{(Depricated) A \code{character} vector of prefecture codes or names.}
}
\value{
A \code{jpcity_pref} object.
}
\description{
Parse prefecture codes or names
}
