% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref.R
\name{pref_code}
\alias{pref_code}
\title{Get prefecture codes}
\usage{
pref_code(city)
}
\arguments{
\item{city}{A \code{jpcity_city} or \code{jpcity_pref} object.}
}
\value{
A integer vector of prefecture codes.
}
\description{
Get prefecture codes
}
