% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref-legacy.R
\name{pref_data_legacy}
\alias{pref_data_legacy}
\title{Get pref data}
\usage{
pref_data_legacy(pref)
}
\arguments{
\item{pref}{A \code{jpcity_pref} object.}
}
\value{
A data frame.
}
\description{
Get pref data
}
