% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref.R
\name{pref_name}
\alias{pref_name}
\title{Get prefecture names}
\usage{
pref_name(city)
}
\arguments{
\item{city}{A \code{jpcity_city} object.}
}
\value{
A \code{character} vector of prefecture names.
}
\description{
Get prefecture names
}
