% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_city.R
\name{parse_city}
\alias{parse_city}
\title{Parse city codes}
\usage{
parse_city(x, when = NULL, na = c("", "NA"), city_code = deprecated())
}
\arguments{
\item{x}{A \code{character} vector of city codes.}

\item{when}{A \code{character} (year, month, and day components) or date-time
object.}

\item{na}{A \code{character} vector to be treated as missing values.}

\item{city_code}{(Deprecated) A \code{character} vector of city codes.}
}
\value{
A \code{jpcity_city} object.
}
\description{
Parse city codes
}
