% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref.R
\name{parse_pref}
\alias{parse_pref}
\title{Parse prefecture names}
\usage{
parse_pref(pref_name)
}
\arguments{
\item{pref_name}{A \code{character} vector of prefecture names.}
}
\value{
A \code{jpcity_pref} object.
}
\description{
Parse prefecture names
}
